chore: :truck: moves App.vue to src folder
